supplier_ingest/requirements.txt
python-telegram-bot~=22.1
Pillow
pytesseract
supabase
requests
rapidfuzz
opencv-python
numpy
easyocr
opencv-python